Fix bug in cycle function
The bug was: it wasn’t working, and I can’t imitate the Haskell
function exactly, because I get a stack overflow without real lazy
evaluation. So the solution is kind of a hack, since I imported much of
the code in from the listRangeLazyBy function, but it works and is
seldom likely to be a useful function anyway?