Remove undefined check from map function
I can’t figure out how to get code coverage for this check, so perhaps
it isn’t needed anymore, or Babel was updated to fix whatever bug was
causing the problem that made me put this check here in the first place
whatever that was. I imagine something to do with function application.